requirements.txt
mysql-connector-python
py4j
pyspark
python-dotenv
setuptools